# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Online Shop API",
      "version": "1.0.0",
      "description": "API Documentation for Online Shop"
    },
    "paths": {
      "/Register": {
        "post": {
          "tags": ["Auth"],
          "summary": "Register a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "default": "user"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User registered successfully"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/Login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Login a user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Invalid credentials"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/categories": {
        "get": {
          "tags": ["Categories"],
          "summary": "Get all categories",
          "responses": {
            "200": {
              "description": "List of categories",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "post": {
          "tags": ["Categories"],
          "summary": "Create a new category",
          "security": [{
            "bearerAuth": []
          }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": ["name"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Category created successfully"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/categories/{id}": {
        "get": {
          "tags": ["Categories"],
          "summary": "Get a category by ID",
          "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }],
          "responses": {
            "200": {
              "description": "Category details",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Category not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "put": {
          "tags": ["Categories"],
          "summary": "Update a category",
          "security": [{
            "bearerAuth": []
          }],
          "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": ["name"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Category updated successfully"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "delete": {
          "tags": ["Categories"],
          "summary": "Delete a category",
          "security": [{
            "bearerAuth": []
          }],
          "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }],
          "responses": {
            "200": {
              "description": "Category deleted successfully"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/products": {
        "get": {
          "tags": ["Products"],
          "summary": "Get all products",
          "responses": {
            "200": {
              "description": "List of products",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        },
                        "categoryId": {
                          "type": "integer"
                        },
                        "image": {
                          "type": "string",
                          "format": "base64"
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "post": {
          "tags": ["Products"],
          "summary": "Create a new product",
          "security": [{
            "bearerAuth": []
          }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "categoryId": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string",
                      "format": "base64"
                    }
                  },
                  "required": ["name", "price", "categoryId"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Product created successfully"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "tags": ["Products"],
          "summary": "Get a product by ID",
          "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }],
          "responses": {
            "200": {
              "description": "Product details",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "categoryId": {
                        "type": "integer"
                      },
                      "image": {
                        "type": "string",
                        "format": "base64"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "put": {
          "tags": ["Products"],
          "summary": "Update a product",
          "security": [{
            "bearerAuth": []
          }],
          "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "categoryId": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string",
                      "format": "base64"
                    }
                  },
                  "required": ["name", "price", "categoryId"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Product updated successfully"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "delete": {
          "tags": ["Products"],
          "summary": "Delete a product",
          "security": [{
            "bearerAuth": []
          }],
          "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }],
          "responses": {
            "200": {
              "description": "Product deleted successfully"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [{
      "bearerAuth": []
    }]
  }